samx5x: support external clock sources
Adds two arguments to clock_init, to specify a frequency for XOSC0, and
if it is a crystal or not. If xosc_freq is non-zero, DPLL0 will use
XOSC0 as its REFCLK on SAM_D5x_E5x series chips. These arguments
currently have no effect for SAMD21 series chips.

This commit also allows osc_get_frequency and dpll_get_frequency to
return valid frequencies for XOSC0 if it is being used to generate the
system clock.

Signed-off-by: Qyriad <qyriad@qyriad.me>